Do not delete a Spy with ongoing missions

+++ app/Actions/Spy/DeleteSpyAction.php
@@ -4,12 +4,21 @@ declare(strict_types=1);
 
 namespace App\Actions\Spy;
 
+use App\Exceptions\SpyDeletionException;
 use App\Models\Spy;
 
 class DeleteSpyAction
 {
+    /**
+     * @throws \Throwable
+     */
     public function handle(Spy $spy): void
     {
+        throw_if(
+            $spy->missions()->active()->count() > 0,
+            new SpyDeletionException('Spy has ongoing missions and cannot be deleted.'),
+        );
+
         $spy->delete();
     }
 }
